fix(cute_dsl): make the optional-dependency guard independent of cutlass
`import flashinfer` hard-required `nvidia-cutlass-dsl`, which is an optional
dependency.

`flashinfer/gemm/kernels/__init__.py` is written as a textbook optional-import
guard:

    from flashinfer.cute_dsl.utils import is_cute_dsl_available
    if is_cute_dsl_available():
        from .bmm_fp8_wrapper import ...

but `flashinfer/cute_dsl/utils.py` did `import cutlass` at module scope. To ask
whether cutlass was installed you first had to import a module that requires
cutlass, so when it was absent the import raised `ModuleNotFoundError` and the
guard was never reached — unreachable in the only circumstance it exists for.
`flashinfer/__init__.py` imports `.gemm` eagerly, so this ran on every import of
the package, not just for consumers using CuTe-DSL paths.

The probes' own docstrings promise graceful degradation ("fall back to a
plain-CUDA implementation", "on an older DSL the rest of the package still
works"), and they are already implemented with `importlib.util.find_spec`, which
needs no cutlass import at all. The hard import was required by other code in the
same file, not by the probes.

Move them to a new `flashinfer/cute_dsl/availability.py` that imports nothing
from cutlass at module scope; the few functions that need the DSL's own metadata
(`is_cute_dsl_arch_supported` and helpers, which read `cutlass.base_dsl.arch`)
already import it lazily inside a `try`. `flashinfer/cute_dsl/__init__.py` now
imports only that module eagerly and moves the cutlass-requiring re-exports
(`make_ptr`, `get_cutlass_dtype`, ...) behind the availability guard, so the
subpackage itself is importable without the DSL. `flashinfer/cute_dsl/utils.py`
re-exports the probes, so existing `from flashinfer.cute_dsl.utils import ...`
call sites keep working when the DSL is installed.

Call sites that run before the DSL is known present now import from
`.availability`. This also fixes the lazy backend predicates in `norm`,
`mla/_core`, `kda_kernels` and `gemm_base`, which were meant to return `False`
when the DSL is absent but would have raised `ModuleNotFoundError` instead.

Two modules had already worked around this by hand-rolling the same
`find_spec("cutlass")` check inline, one of them with a comment naming the cause
("without importing cute_dsl/utils.py, which has top-level `import cutlass`");
both now call the shared probe.

Add a regression test that imports flashinfer with cutlass hidden from the import
system, so this cannot silently return the next time an import is added to the
base path — which is exactly how it arrived.
