+++ flashinfer/mla/_core.py
@@ -2224,7 +2224,7 @@ def _cute_dsl_incompatibility_reason(
     cc = get_compute_capability(query.device)
     if cc[0] < 10:
         return f"cute-dsl backend (MLA decode kernel) requires SM100+, got SM{cc[0]}{cc[1]}"
-    from ..cute_dsl.utils import is_cute_dsl_arch_supported
+    from ..cute_dsl.availability import is_cute_dsl_arch_supported
 
     if not is_cute_dsl_arch_supported(*cc):
         return (
